requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
pyTelegramBotAPI>=4.12.0 